requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.6
Flask==3.1.0
frozenlist==1.5.0
groq==0.14.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instructor==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.59.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyserial==3.5
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typeguard==4.4.1
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
